AttentionAgent/requirements.txt
cma
gin-config
gym==0.15.3
torch
torchvision
opencv-python
# new below
gym3
wandb
procgen
moviepy

# conda install pytorch torchvision torchaudio cudatoolkit=11.1 -c pytorch -c nvidia